Simplify default payout calculation and remove unused class
