+++ packages/apps/reputation-oracle/server/src/modules/escrow-completion/payouts-calculation/index.ts
@@ -1,4 +1,3 @@
 export { CvatPayoutsCalculator } from './cvat-payouts-calculator';
 export { DefaultPayoutsCalculator } from './default-payouts-calculator';
-export { FinalResultsPayoutsCalculator } from './final-results-payouts-calculator';
 export * from './types';
